Struct aws_sdk_personalize::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonPersonalize.

This client allows ergonomic access to a AmazonPersonalize-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_batch_inference_job(&self) -> CreateBatchInferenceJob<C, M, R>

pub fn create_campaign(&self) -> CreateCampaign<C, M, R>

pub fn create_dataset(&self) -> CreateDataset<C, M, R>

pub fn create_dataset_export_job(&self) -> CreateDatasetExportJob<C, M, R>

pub fn create_dataset_group(&self) -> CreateDatasetGroup<C, M, R>

pub fn create_dataset_import_job(&self) -> CreateDatasetImportJob<C, M, R>

pub fn create_event_tracker(&self) -> CreateEventTracker<C, M, R>

pub fn create_filter(&self) -> CreateFilter<C, M, R>

pub fn create_schema(&self) -> CreateSchema<C, M, R>

pub fn create_solution(&self) -> CreateSolution<C, M, R>

pub fn create_solution_version(&self) -> CreateSolutionVersion<C, M, R>

pub fn delete_campaign(&self) -> DeleteCampaign<C, M, R>

pub fn delete_dataset(&self) -> DeleteDataset<C, M, R>

pub fn delete_dataset_group(&self) -> DeleteDatasetGroup<C, M, R>

pub fn delete_event_tracker(&self) -> DeleteEventTracker<C, M, R>

pub fn delete_filter(&self) -> DeleteFilter<C, M, R>

pub fn delete_schema(&self) -> DeleteSchema<C, M, R>

pub fn delete_solution(&self) -> DeleteSolution<C, M, R>

pub fn describe_algorithm(&self) -> DescribeAlgorithm<C, M, R>

pub fn describe_batch_inference_job(&self) -> DescribeBatchInferenceJob<C, M, R>

pub fn describe_campaign(&self) -> DescribeCampaign<C, M, R>

pub fn describe_dataset(&self) -> DescribeDataset<C, M, R>

pub fn describe_dataset_export_job(&self) -> DescribeDatasetExportJob<C, M, R>

pub fn describe_dataset_group(&self) -> DescribeDatasetGroup<C, M, R>

pub fn describe_dataset_import_job(&self) -> DescribeDatasetImportJob<C, M, R>

pub fn describe_event_tracker(&self) -> DescribeEventTracker<C, M, R>

pub fn describe_feature_transformation(
&self
) -> DescribeFeatureTransformation<C, M, R>

pub fn describe_filter(&self) -> DescribeFilter<C, M, R>

pub fn describe_recipe(&self) -> DescribeRecipe<C, M, R>

pub fn describe_schema(&self) -> DescribeSchema<C, M, R>

pub fn describe_solution(&self) -> DescribeSolution<C, M, R>

pub fn describe_solution_version(&self) -> DescribeSolutionVersion<C, M, R>

pub fn get_solution_metrics(&self) -> GetSolutionMetrics<C, M, R>

pub fn list_batch_inference_jobs(&self) -> ListBatchInferenceJobs<C, M, R>

pub fn list_campaigns(&self) -> ListCampaigns<C, M, R>

pub fn list_dataset_export_jobs(&self) -> ListDatasetExportJobs<C, M, R>

pub fn list_dataset_groups(&self) -> ListDatasetGroups<C, M, R>

pub fn list_dataset_import_jobs(&self) -> ListDatasetImportJobs<C, M, R>

pub fn list_datasets(&self) -> ListDatasets<C, M, R>

pub fn list_event_trackers(&self) -> ListEventTrackers<C, M, R>

pub fn list_filters(&self) -> ListFilters<C, M, R>

pub fn list_recipes(&self) -> ListRecipes<C, M, R>

pub fn list_schemas(&self) -> ListSchemas<C, M, R>

pub fn list_solutions(&self) -> ListSolutions<C, M, R>

pub fn list_solution_versions(&self) -> ListSolutionVersions<C, M, R>

pub fn stop_solution_version_creation(
&self
) -> StopSolutionVersionCreation<C, M, R>

pub fn update_campaign(&self) -> UpdateCampaign<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self