Module aws_sdk_personalize::error

Modules

invalid_input_exception

See InvalidInputException

invalid_next_token_exception

See InvalidNextTokenException

limit_exceeded_exception

See LimitExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

Structs

CreateBatchInferenceJobError

CreateCampaignError

CreateDatasetError

CreateDatasetExportJobError

CreateDatasetGroupError

CreateDatasetImportJobError

CreateEventTrackerError

CreateFilterError

CreateSchemaError

CreateSolutionError

CreateSolutionVersionError

DeleteCampaignError

DeleteDatasetError

DeleteDatasetGroupError

DeleteEventTrackerError

DeleteFilterError

DeleteSchemaError

DeleteSolutionError

DescribeAlgorithmError

DescribeBatchInferenceJobError

DescribeCampaignError

DescribeDatasetError

DescribeDatasetExportJobError

DescribeDatasetGroupError

DescribeDatasetImportJobError

DescribeEventTrackerError

DescribeFeatureTransformationError

DescribeFilterError

DescribeRecipeError

DescribeSchemaError

DescribeSolutionError

DescribeSolutionVersionError

GetSolutionMetricsError

InvalidInputException

Provide a valid value for the field or parameter.

InvalidNextTokenException

The token is not valid.

LimitExceededException

The limit on the number of requests per second has been exceeded.

ListBatchInferenceJobsError

ListCampaignsError

ListDatasetExportJobsError

ListDatasetGroupsError

ListDatasetImportJobsError

ListDatasetsError

ListEventTrackersError

ListFiltersError

ListRecipesError

ListSchemasError

ListSolutionVersionsError

ListSolutionsError

ResourceAlreadyExistsException

The specified resource already exists.

ResourceInUseException

The specified resource is in use.

ResourceNotFoundException

Could not find the specified resource.

StopSolutionVersionCreationError

UpdateCampaignError

Enums

CreateBatchInferenceJobErrorKind

CreateCampaignErrorKind

CreateDatasetErrorKind

CreateDatasetExportJobErrorKind

CreateDatasetGroupErrorKind

CreateDatasetImportJobErrorKind

CreateEventTrackerErrorKind

CreateFilterErrorKind

CreateSchemaErrorKind

CreateSolutionErrorKind

CreateSolutionVersionErrorKind

DeleteCampaignErrorKind

DeleteDatasetErrorKind

DeleteDatasetGroupErrorKind

DeleteEventTrackerErrorKind

DeleteFilterErrorKind

DeleteSchemaErrorKind

DeleteSolutionErrorKind

DescribeAlgorithmErrorKind

DescribeBatchInferenceJobErrorKind

DescribeCampaignErrorKind

DescribeDatasetErrorKind

DescribeDatasetExportJobErrorKind

DescribeDatasetGroupErrorKind

DescribeDatasetImportJobErrorKind

DescribeEventTrackerErrorKind

DescribeFeatureTransformationErrorKind

DescribeFilterErrorKind

DescribeRecipeErrorKind

DescribeSchemaErrorKind

DescribeSolutionErrorKind

DescribeSolutionVersionErrorKind

GetSolutionMetricsErrorKind

ListBatchInferenceJobsErrorKind

ListCampaignsErrorKind

ListDatasetExportJobsErrorKind

ListDatasetGroupsErrorKind

ListDatasetImportJobsErrorKind

ListDatasetsErrorKind

ListEventTrackersErrorKind

ListFiltersErrorKind

ListRecipesErrorKind

ListSchemasErrorKind

ListSolutionVersionsErrorKind

ListSolutionsErrorKind

StopSolutionVersionCreationErrorKind

UpdateCampaignErrorKind