Module aws_sdk_personalize::model

Modules

algorithm

See Algorithm

algorithm_image

See AlgorithmImage

auto_ml_config

See AutoMlConfig

auto_ml_result

See AutoMlResult

batch_inference_job

See BatchInferenceJob

batch_inference_job_config

See BatchInferenceJobConfig

batch_inference_job_input

See BatchInferenceJobInput

batch_inference_job_output

See BatchInferenceJobOutput

batch_inference_job_summary

See BatchInferenceJobSummary

campaign

See Campaign

campaign_config

See CampaignConfig

campaign_summary

See CampaignSummary

campaign_update_summary

See CampaignUpdateSummary

categorical_hyper_parameter_range

See CategoricalHyperParameterRange

continuous_hyper_parameter_range

See ContinuousHyperParameterRange

data_source

See DataSource

dataset

See Dataset

dataset_export_job

See DatasetExportJob

dataset_export_job_output

See DatasetExportJobOutput

dataset_export_job_summary

See DatasetExportJobSummary

dataset_group

See DatasetGroup

dataset_group_summary

See DatasetGroupSummary

dataset_import_job

See DatasetImportJob

dataset_import_job_summary

See DatasetImportJobSummary

dataset_schema

See DatasetSchema

dataset_schema_summary

See DatasetSchemaSummary

dataset_summary

See DatasetSummary

default_categorical_hyper_parameter_range

See DefaultCategoricalHyperParameterRange

default_continuous_hyper_parameter_range

See DefaultContinuousHyperParameterRange

default_hyper_parameter_ranges

See DefaultHyperParameterRanges

default_integer_hyper_parameter_range

See DefaultIntegerHyperParameterRange

event_tracker

See EventTracker

event_tracker_summary

See EventTrackerSummary

feature_transformation

See FeatureTransformation

filter

See Filter

filter_summary

See FilterSummary

hpo_config

See HpoConfig

hpo_objective

See HpoObjective

hpo_resource_config

See HpoResourceConfig

hyper_parameter_ranges

See HyperParameterRanges

integer_hyper_parameter_range

See IntegerHyperParameterRange

optimization_objective

See OptimizationObjective

recipe

See Recipe

recipe_summary

See RecipeSummary

s3_data_config

See S3DataConfig

solution

See Solution

solution_config

See SolutionConfig

solution_summary

See SolutionSummary

solution_version

See SolutionVersion

solution_version_summary

See SolutionVersionSummary

tuned_hpo_params

See TunedHpoParams

Structs

Algorithm

Describes a custom algorithm.

AlgorithmImage

Describes an algorithm image.

AutoMlConfig

When the solution performs AutoML (performAutoML is true in

AutoMlResult

When the solution performs AutoML (performAutoML is true in

BatchInferenceJob

Contains information on a batch inference job.

BatchInferenceJobConfig

The configuration details of a batch inference job.

BatchInferenceJobInput

The input configuration of a batch inference job.

BatchInferenceJobOutput

The output configuration parameters of a batch inference job.

BatchInferenceJobSummary

A truncated version of the BatchInferenceJob datatype. The ListBatchInferenceJobs operation returns a list of batch inference job

Campaign

Describes a deployed solution version, otherwise known as a campaign.

CampaignConfig

The configuration details of a campaign.

CampaignSummary

Provides a summary of the properties of a campaign. For a complete listing, call the

CampaignUpdateSummary

Provides a summary of the properties of a campaign update. For a complete listing, call the

CategoricalHyperParameterRange

Provides the name and range of a categorical hyperparameter.

ContinuousHyperParameterRange

Provides the name and range of a continuous hyperparameter.

DataSource

Describes the data source that contains the data to upload to a dataset.

Dataset

Provides metadata for a dataset.

DatasetExportJob

Describes a job that exports a dataset to an Amazon S3 bucket. For more information, see CreateDatasetExportJob.

DatasetExportJobOutput

The output configuration parameters of a dataset export job.

DatasetExportJobSummary

Provides a summary of the properties of a dataset export job. For a complete listing, call the

DatasetGroup

A dataset group is a collection of related datasets (Interactions, User, and Item). You

DatasetGroupSummary

Provides a summary of the properties of a dataset group. For a complete listing, call the

DatasetImportJob

Describes a job that imports training data from a data source (Amazon S3 bucket) to an

DatasetImportJobSummary

Provides a summary of the properties of a dataset import job. For a complete listing, call the

DatasetSchema

Describes the schema for a dataset. For more information on schemas, see

DatasetSchemaSummary

Provides a summary of the properties of a dataset schema. For a complete listing, call the

DatasetSummary

Provides a summary of the properties of a dataset. For a complete listing, call the

DefaultCategoricalHyperParameterRange

Provides the name and default range of a categorical hyperparameter

DefaultContinuousHyperParameterRange

Provides the name and default range of a continuous hyperparameter

DefaultHyperParameterRanges

Specifies the hyperparameters and their default ranges.

DefaultIntegerHyperParameterRange

Provides the name and default range of a integer-valued hyperparameter

EventTracker

Provides information about an event tracker.

EventTrackerSummary

Provides a summary of the properties of an event tracker. For a complete listing, call the

FeatureTransformation

Provides feature transformation information. Feature transformation is the process

Filter

Contains information on a recommendation filter, including its ARN, status, and filter

FilterSummary

A short summary of a filter's attributes.

HpoConfig

Describes the properties for hyperparameter optimization (HPO).

HpoObjective

The metric to optimize during hyperparameter optimization (HPO).

HpoResourceConfig

Describes the resource configuration for hyperparameter optimization (HPO).

HyperParameterRanges

Specifies the hyperparameters and their ranges.

IntegerHyperParameterRange

Provides the name and range of an integer-valued hyperparameter.

OptimizationObjective

Describes the additional objective for the solution, such as maximizing streaming

Recipe

Provides information about a recipe. Each recipe provides an algorithm

RecipeSummary

Provides a summary of the properties of a recipe. For a complete listing, call the

S3DataConfig

The configuration details of an Amazon S3 input or output bucket.

Solution

An object that provides information about a solution. A solution is a trained model

SolutionConfig

Describes the configuration properties for the solution.

SolutionSummary

Provides a summary of the properties of a solution. For a complete listing, call the

SolutionVersion

An object that provides information about a specific version of a Solution.

SolutionVersionSummary

Provides a summary of the properties of a solution version. For a complete listing, call the

TunedHpoParams

If hyperparameter optimization (HPO) was performed, contains the hyperparameter values of

Enums

IngestionMode

ObjectiveSensitivity

RecipeProvider

TrainingMode