Module aws_sdk_personalize::operation

Structs

CreateBatchInferenceJob

Creates a batch inference job. The operation can handle up to 50 million records and the

CreateCampaign

Creates a campaign by deploying a solution version. When a client calls the

CreateDataset

Creates an empty dataset and adds it to the specified dataset group. Use CreateDatasetImportJob to import your training data to a dataset.

CreateDatasetExportJob

CreateDatasetGroup

Creates an empty dataset group. A dataset group contains related datasets that supply data

CreateDatasetImportJob

Creates a job that imports training data from your data source (an Amazon S3 bucket) to an

CreateEventTracker

Creates an event tracker that you use when adding event data to a specified dataset

CreateFilter

Creates a recommendation filter. For more information, see filter.

CreateSchema

Creates an Amazon Personalize schema from the specified schema string. The schema you create

CreateSolution

Creates the configuration for training a model. A trained model is known as

CreateSolutionVersion

Trains or retrains an active solution. A solution is created using the CreateSolution operation and must be in the ACTIVE state before calling

DeleteCampaign

Removes a campaign by deleting the solution deployment. The solution that

DeleteDataset

Deletes a dataset. You can't delete a dataset if an associated

DeleteDatasetGroup

Deletes a dataset group. Before you delete a dataset group, you must delete the

DeleteEventTracker

Deletes the event tracker. Does not delete the event-interactions dataset from

DeleteFilter

Deletes a filter.

DeleteSchema

Deletes a schema. Before deleting a schema, you must delete all

DeleteSolution

Deletes all versions of a solution and the Solution object itself.

DescribeAlgorithm

Describes the given algorithm.

DescribeBatchInferenceJob

Gets the properties of a batch inference job including name, Amazon Resource Name (ARN),

DescribeCampaign

Describes the given campaign, including its status.

DescribeDataset

Describes the given dataset. For more information on datasets, see CreateDataset.

DescribeDatasetExportJob

Describes the dataset export job created by CreateDatasetExportJob,

DescribeDatasetGroup

Describes the given dataset group. For more information on dataset groups, see CreateDatasetGroup.

DescribeDatasetImportJob

Describes the dataset import job created by CreateDatasetImportJob,

DescribeEventTracker

Describes an event tracker. The response includes the trackingId and

DescribeFeatureTransformation

Describes the given feature transformation.

DescribeFilter

Describes a filter's properties.

DescribeRecipe

Describes a recipe.

DescribeSchema

Describes a schema. For more information on schemas, see

DescribeSolution

Describes a solution.

DescribeSolutionVersion

Describes a specific version of a solution. For more information on solutions, see CreateSolution.

GetSolutionMetrics

Gets the metrics for the specified solution version.

ListBatchInferenceJobs

Gets a list of the batch inference jobs that have been performed off of a solution

ListCampaigns

Returns a list of campaigns that use the given solution.

ListDatasetExportJobs

Returns a list of dataset export jobs that use the given dataset. When a dataset is not

ListDatasetGroups

Returns a list of dataset groups. The response provides the properties for each dataset

ListDatasetImportJobs

Returns a list of dataset import jobs that use the given dataset. When a dataset is not

ListDatasets

Returns the list of datasets contained in the given dataset group. The response provides

ListEventTrackers

Returns the list of event trackers associated with the account.

ListFilters

Lists all filters that belong to a given dataset group.

ListRecipes

Returns a list of available recipes. The response provides the properties

ListSchemas

Returns the list of schemas associated with the account. The response provides the

ListSolutionVersions

Returns a list of solution versions for the given solution. When a solution is not

ListSolutions

Returns a list of solutions that use the given dataset group.

StopSolutionVersionCreation

Stops creating a solution version that is in a state of CREATE_PENDING or CREATE IN_PROGRESS.

UpdateCampaign

Updates a campaign by either deploying a new solution or changing the value of the