Struct aws_sdk_personalize::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AmazonPersonalize.

This client allows ergonomic access to a AmazonPersonalize-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn create_batch_inference_job(&self) -> CreateBatchInferenceJob<C, M, R>

Constructs a fluent builder for the CreateBatchInferenceJob operation.

See CreateBatchInferenceJob for more information about the operation and its arguments.

pub fn create_campaign(&self) -> CreateCampaign<C, M, R>

Constructs a fluent builder for the CreateCampaign operation.

See CreateCampaign for more information about the operation and its arguments.

pub fn create_dataset(&self) -> CreateDataset<C, M, R>

Constructs a fluent builder for the CreateDataset operation.

See CreateDataset for more information about the operation and its arguments.

pub fn create_dataset_export_job(&self) -> CreateDatasetExportJob<C, M, R>

Constructs a fluent builder for the CreateDatasetExportJob operation.

See CreateDatasetExportJob for more information about the operation and its arguments.

pub fn create_dataset_group(&self) -> CreateDatasetGroup<C, M, R>

Constructs a fluent builder for the CreateDatasetGroup operation.

See CreateDatasetGroup for more information about the operation and its arguments.

pub fn create_dataset_import_job(&self) -> CreateDatasetImportJob<C, M, R>

Constructs a fluent builder for the CreateDatasetImportJob operation.

See CreateDatasetImportJob for more information about the operation and its arguments.

pub fn create_event_tracker(&self) -> CreateEventTracker<C, M, R>

Constructs a fluent builder for the CreateEventTracker operation.

See CreateEventTracker for more information about the operation and its arguments.

pub fn create_filter(&self) -> CreateFilter<C, M, R>

Constructs a fluent builder for the CreateFilter operation.

See CreateFilter for more information about the operation and its arguments.

pub fn create_schema(&self) -> CreateSchema<C, M, R>

Constructs a fluent builder for the CreateSchema operation.

See CreateSchema for more information about the operation and its arguments.

pub fn create_solution(&self) -> CreateSolution<C, M, R>

Constructs a fluent builder for the CreateSolution operation.

See CreateSolution for more information about the operation and its arguments.

pub fn create_solution_version(&self) -> CreateSolutionVersion<C, M, R>

Constructs a fluent builder for the CreateSolutionVersion operation.

See CreateSolutionVersion for more information about the operation and its arguments.

pub fn delete_campaign(&self) -> DeleteCampaign<C, M, R>

Constructs a fluent builder for the DeleteCampaign operation.

See DeleteCampaign for more information about the operation and its arguments.

pub fn delete_dataset(&self) -> DeleteDataset<C, M, R>

Constructs a fluent builder for the DeleteDataset operation.

See DeleteDataset for more information about the operation and its arguments.

pub fn delete_dataset_group(&self) -> DeleteDatasetGroup<C, M, R>

Constructs a fluent builder for the DeleteDatasetGroup operation.

See DeleteDatasetGroup for more information about the operation and its arguments.

pub fn delete_event_tracker(&self) -> DeleteEventTracker<C, M, R>

Constructs a fluent builder for the DeleteEventTracker operation.

See DeleteEventTracker for more information about the operation and its arguments.

pub fn delete_filter(&self) -> DeleteFilter<C, M, R>

Constructs a fluent builder for the DeleteFilter operation.

See DeleteFilter for more information about the operation and its arguments.

pub fn delete_schema(&self) -> DeleteSchema<C, M, R>

Constructs a fluent builder for the DeleteSchema operation.

See DeleteSchema for more information about the operation and its arguments.

pub fn delete_solution(&self) -> DeleteSolution<C, M, R>

Constructs a fluent builder for the DeleteSolution operation.

See DeleteSolution for more information about the operation and its arguments.

pub fn describe_algorithm(&self) -> DescribeAlgorithm<C, M, R>

Constructs a fluent builder for the DescribeAlgorithm operation.

See DescribeAlgorithm for more information about the operation and its arguments.

pub fn describe_batch_inference_job(&self) -> DescribeBatchInferenceJob<C, M, R>

Constructs a fluent builder for the DescribeBatchInferenceJob operation.

See DescribeBatchInferenceJob for more information about the operation and its arguments.

pub fn describe_campaign(&self) -> DescribeCampaign<C, M, R>

Constructs a fluent builder for the DescribeCampaign operation.

See DescribeCampaign for more information about the operation and its arguments.

pub fn describe_dataset(&self) -> DescribeDataset<C, M, R>

Constructs a fluent builder for the DescribeDataset operation.

See DescribeDataset for more information about the operation and its arguments.

pub fn describe_dataset_export_job(&self) -> DescribeDatasetExportJob<C, M, R>

Constructs a fluent builder for the DescribeDatasetExportJob operation.

See DescribeDatasetExportJob for more information about the operation and its arguments.

pub fn describe_dataset_group(&self) -> DescribeDatasetGroup<C, M, R>

Constructs a fluent builder for the DescribeDatasetGroup operation.

See DescribeDatasetGroup for more information about the operation and its arguments.

pub fn describe_dataset_import_job(&self) -> DescribeDatasetImportJob<C, M, R>

Constructs a fluent builder for the DescribeDatasetImportJob operation.

See DescribeDatasetImportJob for more information about the operation and its arguments.

pub fn describe_event_tracker(&self) -> DescribeEventTracker<C, M, R>

Constructs a fluent builder for the DescribeEventTracker operation.

See DescribeEventTracker for more information about the operation and its arguments.

pub fn describe_feature_transformation(
&self
) -> DescribeFeatureTransformation<C, M, R>

Constructs a fluent builder for the DescribeFeatureTransformation operation.

See DescribeFeatureTransformation for more information about the operation and its arguments.

pub fn describe_filter(&self) -> DescribeFilter<C, M, R>

Constructs a fluent builder for the DescribeFilter operation.

See DescribeFilter for more information about the operation and its arguments.

pub fn describe_recipe(&self) -> DescribeRecipe<C, M, R>

Constructs a fluent builder for the DescribeRecipe operation.

See DescribeRecipe for more information about the operation and its arguments.

pub fn describe_schema(&self) -> DescribeSchema<C, M, R>

Constructs a fluent builder for the DescribeSchema operation.

See DescribeSchema for more information about the operation and its arguments.

pub fn describe_solution(&self) -> DescribeSolution<C, M, R>

Constructs a fluent builder for the DescribeSolution operation.

See DescribeSolution for more information about the operation and its arguments.

pub fn describe_solution_version(&self) -> DescribeSolutionVersion<C, M, R>

Constructs a fluent builder for the DescribeSolutionVersion operation.

See DescribeSolutionVersion for more information about the operation and its arguments.

pub fn get_solution_metrics(&self) -> GetSolutionMetrics<C, M, R>

Constructs a fluent builder for the GetSolutionMetrics operation.

See GetSolutionMetrics for more information about the operation and its arguments.

pub fn list_batch_inference_jobs(&self) -> ListBatchInferenceJobs<C, M, R>

Constructs a fluent builder for the ListBatchInferenceJobs operation.

See ListBatchInferenceJobs for more information about the operation and its arguments.

pub fn list_campaigns(&self) -> ListCampaigns<C, M, R>

Constructs a fluent builder for the ListCampaigns operation.

See ListCampaigns for more information about the operation and its arguments.

pub fn list_dataset_export_jobs(&self) -> ListDatasetExportJobs<C, M, R>

Constructs a fluent builder for the ListDatasetExportJobs operation.

See ListDatasetExportJobs for more information about the operation and its arguments.

pub fn list_dataset_groups(&self) -> ListDatasetGroups<C, M, R>

Constructs a fluent builder for the ListDatasetGroups operation.

See ListDatasetGroups for more information about the operation and its arguments.

pub fn list_dataset_import_jobs(&self) -> ListDatasetImportJobs<C, M, R>

Constructs a fluent builder for the ListDatasetImportJobs operation.

See ListDatasetImportJobs for more information about the operation and its arguments.

pub fn list_datasets(&self) -> ListDatasets<C, M, R>

Constructs a fluent builder for the ListDatasets operation.

See ListDatasets for more information about the operation and its arguments.

pub fn list_event_trackers(&self) -> ListEventTrackers<C, M, R>

Constructs a fluent builder for the ListEventTrackers operation.

See ListEventTrackers for more information about the operation and its arguments.

pub fn list_filters(&self) -> ListFilters<C, M, R>

Constructs a fluent builder for the ListFilters operation.

See ListFilters for more information about the operation and its arguments.

pub fn list_recipes(&self) -> ListRecipes<C, M, R>

Constructs a fluent builder for the ListRecipes operation.

See ListRecipes for more information about the operation and its arguments.

pub fn list_schemas(&self) -> ListSchemas<C, M, R>

Constructs a fluent builder for the ListSchemas operation.

See ListSchemas for more information about the operation and its arguments.

pub fn list_solutions(&self) -> ListSolutions<C, M, R>

Constructs a fluent builder for the ListSolutions operation.

See ListSolutions for more information about the operation and its arguments.

pub fn list_solution_versions(&self) -> ListSolutionVersions<C, M, R>

Constructs a fluent builder for the ListSolutionVersions operation.

See ListSolutionVersions for more information about the operation and its arguments.

pub fn stop_solution_version_creation(
&self
) -> StopSolutionVersionCreation<C, M, R>

Constructs a fluent builder for the StopSolutionVersionCreation operation.

See StopSolutionVersionCreation for more information about the operation and its arguments.

pub fn update_campaign(&self) -> UpdateCampaign<C, M, R>

Constructs a fluent builder for the UpdateCampaign operation.

See UpdateCampaign for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.